[REFACTOR] Divide MainPage Lists to Atoms and Setting ReactQuery Default Options
메인페이지 일부 Atoms으로 분리 및 리액트쿼리 기본 설정 추가
